Thu, May 4, 2023, 7:59 PM -04:00

+++ Week 4: Memory - PSET & Lab/filter-less/helpers.c
@@ -12,7 +12,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width; j++)
         {
             //grayscale is defined by averaging R,G, and B colors, and changing every pixel hex color to the new gray hex value.
-            grayscaleHex = round((float)((image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3));
+            grayscaleHex = round((image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3);
             image[i][j].rgbtRed = grayscaleHex;
             image[i][j].rgbtGreen = grayscaleHex;
             image[i][j].rgbtBlue = grayscaleHex;
